eliminar notificaciones locale al abrir la app

+++ src/app/pages/public-profile/menu/menu.component.ts
@@ -33,7 +33,7 @@ export class MenuComponent implements OnInit {
           description: 'WINK.DIALOGUES.MESSAGES.DELETE_WINK',
         }).subscribe(
           async (resp: any) => {
-            if (resp.value) {
+            if (resp && resp.value) {
               this.popoverController.dismiss();
               await this.winkService.DeleteWink(this.wink);
             }
